requirements.txt
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
dj-database-url==3.1.2
django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mistune==3.2.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
python-frontmatter==1.1.0
pyyaml==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
sqlparse==0.5.5
typing-extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
